backend/requirements.txt
Django>=4.2
djangorestframework
django-filter
django-cors-headers
mysqlclient
python-dotenv
gunicorn
